Pause-resume patch fixes

+++ gillespy2/solvers/utilities/solverutils.py
@@ -91,19 +91,19 @@ def update_species_init_values(listOfSpecies, species, variables, resume = None)
     populations = ''
     for i in range(len(species) - 1):
         if species[i] in variables:
-            populations += '{} '.format(int(variables[species[i]]))
+            populations += '{} '.format(float(variables[species[i]]))
         else:
             if resume is not None:
-                populations += '{} '.format(int(resume[species[i]][-1]))
+                populations += '{} '.format(float(resume[species[i]][-1]))
             else:
-                populations += '{} '.format(int(listOfSpecies[species[i]].initial_value))
+                populations += '{} '.format(float(listOfSpecies[species[i]].initial_value))
     if species[-1] in variables:
-        populations += '{}'.format(int(variables[species[-1]]))
+        populations += '{}'.format(float(variables[species[-1]]))
     else:
         if resume is not None:
-            populations += '{} '.format(int(resume[species[-1]][-1]))
+            populations += '{} '.format(float(resume[species[-1]][-1]))
         else:
-            populations += '{}'.format(int(listOfSpecies[species[-1]].initial_value))
+            populations += '{}'.format(float(listOfSpecies[species[-1]].initial_value))
     return populations
 
 def change_param_values(listOfParameters, parameters, volume, variables):
